+++ lib/ui/groups/GroupPostDetailPanel.dart
@@ -136,63 +136,48 @@ class _GroupPostDetailPanelState extends State<GroupPostDetailPanel> implements 
           _buildRepliesSection(),
           _buildPostEdit(),
           ],)),
-       Container(
-                key: _sliverHeaderKey,
-                color: Styles().colors!.background,
-                padding: EdgeInsets.only(left: _outerPadding, bottom: 3),
-                child: Column(
-                    crossAxisAlignment: CrossAxisAlignment.start,
-                    mainAxisSize: MainAxisSize.min,
-                    children: [
-                      Row(
-                          mainAxisAlignment: MainAxisAlignment.start,
-                          crossAxisAlignment: CrossAxisAlignment.center,
-                          children: [
-                            Expanded(
-                                child: Semantics(
-                                    sortKey: OrdinalSortKey(1),
-                                    container: true,
-                                    child: Text(
-                                        StringUtils.ensureNotEmpty(_post?.subject),
-                                        maxLines: 5,
-                                        overflow: TextOverflow.ellipsis,
-                                        style: TextStyle(
-                                            fontFamily:
-                                            Styles().fontFamilies!.bold,
-                                            fontSize: 24,
-                                            color: Styles()
-                                                .colors!
-                                                .fillColorPrimary)))),
-                            Visibility(visible: _isEditPostVisible && !widget.hidePostOptions, child:
-                              Semantics(container: true, sortKey: OrdinalSortKey(5), child:
-                                Container(child:
-                                  Semantics(label: Localization().getStringEx('panel.group.detail.post.reply.edit.label', "Edit"), button: true, child:
-                                    GestureDetector(onTap: _onTapEditMainPost, child:
-                                      Padding(padding: EdgeInsets.only(left: 8, top: 22, bottom: 10, right: 8), child:
-                                        Image.asset('images/icon-edit.png', width: 18, height: 18, excludeFromSemantics: true,))))))),
+          Container(key: _sliverHeaderKey, color: Styles().colors!.background, padding: EdgeInsets.only(left: _outerPadding, bottom: 3), child:
+            Column(crossAxisAlignment: CrossAxisAlignment.start, mainAxisSize: MainAxisSize.min, children: [
+              Row(children: [
+                Expanded(child:
+                  Semantics(sortKey: OrdinalSortKey(1), container: true, child:
+                    Text(StringUtils.ensureNotEmpty(_post?.subject), maxLines: 5, overflow: TextOverflow.ellipsis, style:
+                      TextStyle(fontFamily: Styles().fontFamilies!.bold, fontSize: 24, color: Styles().colors!.fillColorPrimary)
+                    )
+                  )
+                ),
 
-                            Visibility(visible: _isDeletePostVisible && !widget.hidePostOptions, child:
-                              Semantics(container: true, sortKey: OrdinalSortKey(5), child:
-                                Container(child:
-                                  Semantics(label: Localization().getStringEx('panel.group.detail.post.reply.delete.label', "Delete"), button: true, child:
-                                    GestureDetector(onTap: _onTapDeletePost, child:
-                                        Padding(padding: EdgeInsets.only(left: 8, top: 22, bottom: 10, right: 8), child:
-                                          Image.asset('images/trash.png', width: 18, height: 18, excludeFromSemantics: true,))))))),
+                Visibility(visible: _isEditPostVisible && !widget.hidePostOptions, child:
+                  Semantics(container: true, sortKey: OrdinalSortKey(5), child:
+                    Container(child:
+                      Semantics(label: Localization().getStringEx('panel.group.detail.post.reply.edit.label', "Edit"), button: true, child:
+                        GestureDetector(onTap: _onTapEditMainPost, child:
+                          Padding(padding: EdgeInsets.only(left: 8, top: 22, bottom: 10, right: 8), child:
+                            Image.asset('images/icon-edit.png', width: 18, height: 18, excludeFromSemantics: true,))))))),
 
-                            Visibility(visible: _isReplyVisible && !widget.hidePostOptions, child:
-                              Semantics(label: Localization().getStringEx('panel.group.detail.post.button.report.label', "Report"), button: true, child:
-                                GestureDetector( onTap: _onTapHeaderReportAbuse, child:
-                                    Padding(padding: EdgeInsets.only(left: 8, top: 22, bottom: 10, right: 8), child:
-                                      Image.asset('images/icon-feedback.png', width: 18, height: 18, fit: BoxFit.fill, excludeFromSemantics: true,))))),
+                Visibility(visible: _isDeletePostVisible && !widget.hidePostOptions, child:
+                  Semantics(container: true, sortKey: OrdinalSortKey(5), child:
+                    Container(child:
+                      Semantics(label: Localization().getStringEx('panel.group.detail.post.reply.delete.label', "Delete"), button: true, child:
+                        GestureDetector(onTap: _onTapDeletePost, child:
+                            Padding(padding: EdgeInsets.only(left: 8, top: 22, bottom: 10, right: 8), child:
+                              Image.asset('images/trash.png', width: 18, height: 18, excludeFromSemantics: true,))))))),
 
-                            Visibility(visible: _isReportAbuseVisible && !widget.hidePostOptions, child:
-                              Semantics(label: Localization().getStringEx('panel.group.detail.post.reply.reply.label', "Reply"), button: true, child:
-                                GestureDetector(onTap: _onTapHeaderReply, child:
-                                    Padding(padding: EdgeInsets.only(left: 8, top: 22, bottom: 10, right: 16), child:
-                                      Image.asset('images/icon-group-post-reply.png', width: 18, height: 18, fit: BoxFit.fill, excludeFromSemantics: true,))))),
+                Visibility(visible: _isReportAbuseVisible && !widget.hidePostOptions, child:
+                  Semantics(label: Localization().getStringEx('panel.group.detail.post.button.report.label', "Report"), button: true, child:
+                    GestureDetector( onTap: () => _onTapReportAbusePostOptions(), child:
+                        Padding(padding: EdgeInsets.only(left: 8, top: 22, bottom: 10, right: 8), child:
+                          Image.asset('images/icon-feedback.png', width: 18, height: 18, fit: BoxFit.fill, excludeFromSemantics: true,))))),
 
-                          ]),
-                    ]))
+                Visibility(visible: _isReplyVisible && !widget.hidePostOptions, child:
+                  Semantics(label: Localization().getStringEx('panel.group.detail.post.reply.reply.label', "Reply"), button: true, child:
+                    GestureDetector(onTap: _onTapHeaderReply, child:
+                        Padding(padding: EdgeInsets.only(left: 8, top: 22, bottom: 10, right: 16), child:
+                          Image.asset('images/icon-group-post-reply.png', width: 18, height: 18, fit: BoxFit.fill, excludeFromSemantics: true,))))),
+
+              ]),
+            ])
+          )
       ]),
       Visibility(
           visible: _loading,
@@ -564,6 +549,36 @@ class _GroupPostDetailPanelState extends State<GroupPostDetailPanel> implements 
     });
   }
 
+  void _onTapReportAbusePostOptions() {
+    Analytics().logSelect(target: 'Post Options');
+    showModalBottomSheet(
+      context: context,
+      backgroundColor: Styles().colors!.white,
+      isScrollControlled: true,
+      isDismissible: true,
+      shape: RoundedRectangleBorder(borderRadius: BorderRadius.vertical(top: Radius.circular(24)),),
+      builder: (context) {
+        return Container(
+          padding: EdgeInsets.symmetric(horizontal: 16, vertical: 17),
+          child: Column(
+            mainAxisSize: MainAxisSize.min,
+            children: <Widget>[
+              Visibility(visible: _isReportAbuseVisible, child: RibbonButton(
+                leftIconAsset: "images/icon-feedback.png",
+                label: Localization().getStringEx("panel.group.detail.post.button.report.students_dean.labe", "Report to Dean of Students"),
+                onTap: () => _onTapReportAbuse(type: GroupPostReportAbuseType.deanOfStudents, post: widget.post),
+              )),
+              Visibility(visible: _isReportAbuseVisible, child: RibbonButton(
+                leftIconAsset: "images/icon-feedback.png",
+                label: Localization().getStringEx("panel.group.detail.post.button.report.group_admins.labe", "Report to Group Administrator(s)"),
+                onTap: () => _onTapReportAbuse(type: GroupPostReportAbuseType.groupAdmins, post: widget.post),
+              )),
+            ],
+          ),
+        );
+      });
+  }
+
   void _onTapReplyOptions(GroupPost? reply) {
     Analytics().logSelect(target: 'Reply Options');
     showModalBottomSheet(
@@ -604,11 +619,13 @@ class _GroupPostDetailPanelState extends State<GroupPostDetailPanel> implements 
               )),
               Visibility(visible: _isReportAbuseVisible, child: RibbonButton(
                 leftIconAsset: "images/icon-feedback.png",
-                label: Localization().getStringEx("panel.group.detail.post.button.report.label", "Report"),
-                onTap: () {
-                  Navigator.of(context).pop();
-                  _onTapMenuReportAbuse(reply);
-                },
+                label: Localization().getStringEx("panel.group.detail.post.button.report.students_dean.labe", "Report to Dean of Students"),
+                onTap: () => _onTapReportAbuse(type: GroupPostReportAbuseType.deanOfStudents, post: reply),
+              )),
+              Visibility(visible: _isReportAbuseVisible, child: RibbonButton(
+                leftIconAsset: "images/icon-feedback.png",
+                label: Localization().getStringEx("panel.group.detail.post.button.report.group_admins.labe", "Report to Group Administrator(s)"),
+                onTap: () => _onTapReportAbuse(type: GroupPostReportAbuseType.groupAdmins, post: reply),
               )),
             ],
           ),
@@ -674,16 +691,16 @@ class _GroupPostDetailPanelState extends State<GroupPostDetailPanel> implements 
     _scrollToPostEdit();
   }
 
-  void _onTapHeaderReportAbuse() {
-    Analytics().logSelect(target: 'Report Abuse');
-    Navigator.of(context).push(CupertinoPageRoute(builder: (context) => GroupPostReportAbuse(groupId: widget.group?.id, postId: widget.post?.id)));
+  void _onTapReportAbuse({required GroupPostReportAbuseType type, GroupPost? post}) {
+    String anapyticsSarget;
+    switch(type) {
+      case GroupPostReportAbuseType.deanOfStudents: anapyticsSarget = 'Report Abuse to Dean of Students'; break;
+      case GroupPostReportAbuseType.groupAdmins: anapyticsSarget = 'Report Abuse to Group Administrator(s)'; break;
+    }
+    Analytics().logSelect(target: anapyticsSarget);
+    Navigator.of(context).pushReplacement(CupertinoPageRoute(builder: (context) => GroupPostReportAbuse(type: type, groupId: widget.group?.id, postId: (post ?? widget.post)?.id)));
   }
 
-  void _onTapMenuReportAbuse(GroupPost? post) {
-    Analytics().logSelect(target: 'Report Abuse');
-    Navigator.of(context).pushReplacement(CupertinoPageRoute(builder: (context) => GroupPostReportAbuse(groupId: widget.group?.id, postId: post?.id)));
-  }
-  
   void _onTapEditMainPost(){
     _mainPostUpdateData = PostDataModel(body:_post?.body, imageUrl: _post?.imageUrl, members: GroupMembersSelectionWidget.constructUpdatedMembersList(selection:_post?.members, upToDateMembers: _allMembersAllowedToPost));
     if(mounted){